rename pitstop helper to pressure helper
