+++ Race_Element.HUD.ACC/Overlays/Driving/PressureHelper/PressureHelperConfiguration.cs
@@ -1,13 +1,8 @@
 ﻿using RaceElement.HUD.Overlay.Configuration;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace RaceElement.HUD.ACC.Overlays.Driving.PitstopHelper
+namespace RaceElement.HUD.ACC.Overlays.Driving.PressureHelper
 {
-    internal sealed class PitstopHelperConfiguration : OverlayConfiguration
+    internal sealed class PressureHelperConfiguration : OverlayConfiguration
     {
         [ConfigGrouping("Tyre Pressures", "Adjust tyre pressure targets to be used in the hud.")]
         public TyrePressures Pressures { get; init; } = new();
